requirements.txt
Django==1.11.10
wagtail>=2.0,<2.1
beautifulsoup4==4.6.0
boto3==1.6.3
botocore==1.9.3
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.2
dj-static==0.0.6
django-modelcluster==3.1
django-storages==1.6.5
django-taggit==0.22.2
django-toolbelt==0.0.1
django-treebeard==4.2.0
djangorestframework==3.6.4
docutils==0.14
gunicorn==19.7.1
html5lib==0.999999999
idna==2.6
jmespath==0.9.3
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
static3==0.7.0
Unidecode==1.0.22
urllib3==1.22
wagtail==2.0
webencodings==0.5.1
whitenoise==3.3.1
